refactor: code in fs modules

+++ src/commands/fs/mv.js
@@ -1,28 +1,35 @@
 import { createReadStream, createWriteStream, unlink } from 'node:fs';
-import { basename, isAbsolute, join, dirname } from 'node:path';
+import { basename, join } from 'node:path';
 import { cwd } from 'node:process';
-import { fileURLToPath } from 'node:url';
+
+import { normalizePath } from '../../utils/normalizePath.js';
+import { ERRORS } from '../../utils/constants.js';
 
 export const mv = (params) => {
   try {
-    let path = params.split(' ')[0];
-    const newPath = join(params.split(' ')[1], basename(path));
+    const path = normalizePath(params.split(' ')[0]);
+    const newPath = normalizePath(join(params.split(' ')[1], basename(path)));
+
+    if (path === newPath) {
+      console.log('Such a file already exists in the specified directory')
+      throw new Error();
+    }
 
     const readStream = createReadStream(path, 'utf-8');
     const writeStream = createWriteStream(newPath);
 
-    readStream.on('error', (err) => console.error('Enter the correct path'));
-    writeStream.on('error', (err) => console.error('Enter the correct path'));
+    readStream.on('error', (err) => console.error(ERRORS.OPERATION_FAILED));
+    writeStream.on('error', (err) => console.error(ERRORS.OPERATION_FAILED));
 
     writeStream.on('finish', () => {
       unlink(path, (err) => {
-        if (err) console.error(err);
+        if (err) console.error(ERRORS.OPERATION_FAILED);
         console.log(`You are currently in ${cwd()}`)
       })
     })
     readStream.pipe(writeStream);
 
   } catch (err) {
-    console.error('Enter the correct path');
+    console.error(ERRORS.INVALID_INPUT);
   }
 }
